Salas e items en el online

+++ Rogue-Lite/Assets/Game/Scripts/Player/PlayerStatus.cs
@@ -4,25 +4,25 @@ using UnityEngine;
 
 public class PlayerStatus : MonoBehaviour
 {
-    public float health = 100;
-    public float damage = 18;
+    public int health = 100;
+    public int damage = 18; // Por implementar
     public float movementSpeed = 1;
-    public int goldPerEnemy;
-    public int gold = 0;
-    public float healthSteal = 0.0f;
+    public int goldPerEnemy; // Por implementar
+    public int gold = 0; // Por implementar
+    public float healthSteal = 0.0f; // Por implementar
     // public float attackSpeed = 1.0f;
-    public bool hasHalbert = false;
-    public bool hasRapier = false;
-    public bool hasSpecialSkillOverpowered = false;
-    public bool hasMainAttackOverpowered = false;
-    public float luck = 1;
-    public bool enemiesExplodes = false;
-    public float agility = 1;
+    public bool hasHalbert = false; // Por implementar (que no aparezca otra arma después)
+    public bool hasRapier = false; // Por implementar (que no aparezca otra arma después)
+    public bool hasSpecialSkillOverpowered = false; // Por implementar
+    public bool hasMainAttackOverpowered = false; // Por implementar
+    public float luck = 1; // Por implementar
+    public bool enemiesExplodes = false; // Por implementar
+    public float agility = 1; // Por implementar
     public bool survivesToLetalAttack = false;
-    public float lessInitialLifeForEnemies = 0;
-    public float heal = 20;
-    public float push = 1.0f;
-    public bool spawnGrenadeWhenRolls = false;
-    public float enemiesThreshold = 0.0f;
+    public float lessInitialLifeForEnemies = 0; // Por implementar
+    public float heal = 20; // Por implementar
+    public float push = 1.0f; // Por implementar
+    public bool spawnGrenadeWhenRolls = false; // Por implementar
+    public float enemiesThreshold = 0.0f; // Por implementar
     public bool canRoll = true;
 }
